Rename current test file to reflect what it's actually doing
